+++ test/from-promise.js
@@ -10,10 +10,9 @@ test('test from-promise', t => {
   test('resolves', t => {
     const p = new Promise(resolve => resolve(7));
 
-    const obs = utils.fromPromise(p, 3);
-    t.equal(obs.value, 3);
+    const obs = utils.fromPromise(p);
+    t.equal(obs.value, undefined);
     t.equal(obs.state, 'pending');
-    t.equal(obs.reason, undefined);
     t.ok(obs.promise === p);
 
     mobx.when(
@@ -30,11 +29,24 @@ test('test from-promise', t => {
   test('resolves value', t => {
     const p = new Promise(resolve => resolve(7));
 
-    const obs = utils.fromPromise(p, 3);
-    t.equal(obs.value, 3);
+    const obs = utils.fromPromise(p);
+    t.equal(obs.value, undefined);
     t.equal(obs.state, 'pending');
-    t.equal(obs.reason, undefined);
     t.ok(obs.promise === p);
+
+    mobx.when(
+      () => obs.value === 7,
+      () => {
+        t.equal(obs.state, utils.FULFILLED);
+        t.end();
+      }
+    );
+  });
+
+test('resolves value from promise function', t => {
+    const obs = utils.fromPromise(resolve => resolve(7));
+    t.equal(obs.value, undefined);
+    t.equal(obs.state, 'pending');
 
     mobx.when(
       () => obs.value === 7,
@@ -50,10 +62,9 @@ test('test from-promise', t => {
       reject(7);
     });
 
-    const obs = utils.fromPromise(p, 3);
-    t.equal(obs.value, 3);
+    const obs = utils.fromPromise(p);
+    t.equal(obs.value, undefined);
     t.equal(obs.state, 'pending');
-    t.equal(obs.reason, undefined);
     t.ok(obs.promise === p);
 
     mobx.when(
@@ -61,7 +72,23 @@ test('test from-promise', t => {
       () => {
         t.equal(obs.state, utils.REJECTED);
         t.equal(obs.value, 7);
-        t.equal(obs.reason, 7);
+        t.end();
+      }
+    );
+  });
+
+  test('rejects with reason value from fn', t => {
+    const obs = utils.fromPromise((resolve, reject) => {
+      reject(7);
+    });
+    t.equal(obs.value, undefined);
+    t.equal(obs.state, 'pending');
+
+    mobx.when(
+      () => obs.state !== utils.PENDING,
+      () => {
+        t.equal(obs.state, utils.REJECTED);
+        t.equal(obs.value, 7);
         t.end();
       }
     );
@@ -72,10 +99,9 @@ test('test from-promise', t => {
       throw 7;
     });
 
-    const obs = utils.fromPromise(p, 3);
-    t.equal(obs.value, 3);
+    const obs = utils.fromPromise(p);
+    t.equal(obs.value, undefined);
     t.equal(obs.state, 'pending');
-    t.equal(obs.reason, undefined);
     t.ok(obs.promise === p);
 
     mobx.when(
@@ -83,7 +109,6 @@ test('test from-promise', t => {
       () => {
         t.equal(obs.state, 'rejected');
         t.equal(obs.value, 7);
-        t.equal(obs.reason, 7);
         t.end();
       }
     );
@@ -151,18 +176,109 @@ test('test from-promise', t => {
       }
     );
   });
-  
+
   test('isPromiseBasedObservable, true', t => {
     const obs = utils.fromPromise(Promise.resolve(123));
     t.ok(utils.isPromiseBasedObservable(obs));
     t.end();
   });
-  
+
   test('isPromiseBasedObservable, false', t => {
     t.notOk(utils.isPromiseBasedObservable({}));
     t.end();
   });
-  
+
+  test('state and value are observable, #56', t => {
+    const obs = utils.fromPromise(Promise.resolve(123));
+    t.ok(mobx.isObservable(obs))
+    t.ok(mobx.isObservable(obs, "state"))
+    t.ok(mobx.isObservable(obs, "value"))
+    t.end()
+  })
+
+  test('the resolved value of a promise is not convertd to some deep observable, #54', t => {
+    const someObject = { a: 3 }
+    const obs = utils.fromPromise(Promise.resolve(someObject));
+    obs.promise.then((v) => {
+      t.is(obs.state, utils.FULFILLED)
+      t.false(mobx.isObservable(obs.value))
+      t.true(obs.value === someObject)
+      t.true(v === someObject)
+      t.end()
+    })
+  })
+
+  test('it is possible to create a promise in a rejected state, #36', t=> {
+    const someObject = { a: 3 }
+    const obs = utils.fromPromise.reject(someObject);
+    t.is(obs.state, utils.REJECTED);
+    t.is(obs.value, someObject);
+
+    // still a real promise backing it, which can be thenned...
+    obs.promise.catch((v) => {
+      t.is(obs.state, utils.REJECTED)
+      t.false(mobx.isObservable(obs.value))
+      t.true(obs.value === someObject)
+      t.true(v === someObject)
+      t.end()
+    })
+  })
+
+  test('it is possible to create a promise in a fullfilled state, #36', t=> {
+    const someObject = { a: 3 }
+    const obs = utils.fromPromise.resolve(someObject);
+    t.is(obs.state, utils.FULFILLED);
+    t.is(obs.value, someObject);
+
+    // still a real promise backing it, which can be thenned...
+    obs.promise.then((v) => {
+      t.is(obs.state, utils.FULFILLED)
+      t.false(mobx.isObservable(obs.value))
+      t.true(obs.value === someObject)
+      t.true(v === someObject)
+      t.end()
+    })
+  })
+
+  test('when creating a promise in a fullfilled state it should not fire twice, #36', t=> {
+    let events = 0;
+    const obs = utils.fromPromise.resolve(3);
+
+    mobx.autorun(() => {
+      obs.state; // track state & value
+      obs.value;
+      events++;
+    })
+
+    obs.promise.then((v) => {
+      t.is(events, 1); // only initial run should have run
+      t.end()
+    })
+  })
+
+  test('it creates a real promise, #45', t => {
+    Promise.all([
+      utils.fromPromise.resolve(2),
+      utils.fromPromise(Promise.resolve(3))
+    ]).then(x => {
+      t.deepEqual(x, [2, 3])
+      t.end()
+    })
+  })
+
+  test('it can construct new promises from function, #45', t => {
+    Promise.all([
+      utils.fromPromise((resolve, reject) => {
+        setTimeout(() => resolve(2), 200)
+      }),
+      utils.fromPromise(Promise.resolve(3))
+    ]).then(x => {
+      t.deepEqual(x, [2, 3])
+      t.end()
+    })
+  })
+
+
   t.end();
 });